chore: api라우트 분리 및 프롬프트 수정
