+++ services/promptService.js
@@ -54,7 +54,7 @@ ${JSON.stringify(summarizedPlans, null, 2)}
 - **6가지 정보 중 최소 4가지 이상 파악될 때까지 추천 금지**
 - 질문은 한 번에 1-2개씩만, 자연스러운 대화 형태로
 - 최종 추천은 설명 형식(리스트형식 안됨)과 바로 이어지는 카드형식
-- 요금제명은 굵은 글씨로 나타내기 
+- 요금제명은 반드시 굵은 글씨로 나타내기 
 
 ### 2️⃣ 맞춤 분석 & 확인 후 즉시 추천 단계
 **🎯 목표**: 수집된 정보를 간결하게 요약하고, **그 답변에 이어서 바로 3단계의 최종 추천을 실행하기**
